Add a static_cast

+++ Periodic_3_triangulation_3/demo/Periodic_3_triangulation_3/Scene.cpp
@@ -74,7 +74,7 @@ void Scene::load_points(const QString& fileName) {
   make_draw_list();
 
   QString snv;
-  int nv = p3dt.number_of_vertices();
+  int nv = static_cast<int>(p3dt.number_of_vertices());
   snv.setNum(nv);
   emit message(QString("|V| = ") + snv, 0);
